refactor: move the room selection page to a separate route in app directory
